[JaBi] fix: fixed bug in damaged calculation

+++ src/PokemonDataReader.java
@@ -68,16 +68,16 @@ public class PokemonDataReader {
         return moves;
     }
 
-    public static Map<String, Map<String,Integer>> createTypeTable(String filePath) {
-        Map<String, Map<String,Integer>> moves = new HashMap<>();
+    public static Map<String, Map<String,Double>> createTypeTable(String filePath) {
+        Map<String, Map<String,Double>> moves = new HashMap<>();
         try {
             String jsonData = new String(Files.readAllBytes(Paths.get(filePath)));
             JSONObject types = new JSONObject(jsonData);
             for (String typeName : types.keySet()) {
                 JSONObject typeData = types.getJSONObject(typeName);
-                Map<String,Integer> movesDamage = new HashMap<>();
+                Map<String,Double> movesDamage = new HashMap<>();
                 for (String typeDamageName : typeData.keySet()) {
-                    movesDamage.put(typeDamageName, typeData.getInt(typeDamageName));
+                    movesDamage.put(typeDamageName, typeData.getDouble(typeDamageName));
                 }
                 moves.put(typeName, movesDamage);
             }
